fix: use returned name from response to github create call, rather than submitted name
Github rewrites repository names to substitue a dash for consecutive whitespace, removes special
characters like colon, etc. so after creating the repository, we have to use the new, rewritten name
that Github returns in it's response, rather than the name that we submitted as part of the create
call, otherwise we refer to the new repository by the wrong name in subsequent calls.